[refactor]: 프로젝트 전체 타입 시스템 통합
- EisenhowerBase 공통 타입을 정의하여 중복 코드 제거
- 도메인별 타입(Task, Pomodoro, MindMap)을 공통 타입 기반으로 재구성
- ID 필드 타입을 백엔드 Swagger 명세에 맞춰 number로 통일
- 중복 타입 정의 제거(TodoType → ActualTaskType)
